backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
pydantic-settings>=2.2,<3.0
python-dotenv>=1.0,<2.0
sqlalchemy>=2.0,<3.0
httpx>=0.27,<1.0
pdfplumber>=0.11,<1.0
openai>=1.50,<2.0
python-multipart>=0.0.9,<1.0
